requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
better-profanity==0.7.0
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-celery-beat==2.8.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-nested-admin==4.1.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
greenlet==3.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
openpyxl==3.1.5
pillow==11.1.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
reportlab==4.3.1
requests==2.32.3
service-identity==24.2.0
setuptools==79.0.0
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
